Upgraded TOML messed around w\ dialog, its still crashy

+++ common/src/commonMain/kotlin/com/darkrockstudios/apps/hammer/common/data/encyclopediarepository/EncyclopediaRepository.kt
@@ -78,7 +78,7 @@ abstract class EncyclopediaRepository(
 
 		fun getEntryFilename(entry: EntryContent): String =
 			getEntryFilename(
-				id = entry.id.toInt(),
+				id = entry.id,
 				type = entry.type,
 				name = entry.name
 			)
